requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

pydantic==2.6.1
pydantic-settings==2.2.1

python-dotenv==1.0.1
pyyaml==6.0.1

celery==5.3.6
redis==5.0.1

httpx==0.26.0
requests==2.31.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5

# Utilities
loguru==0.7.2
